Updating flags for kuenm_cal_swd command.
I updated the flags in prepare_swd and kuenm_cal_swd to allow those
functions to run properly. I also re-wrote the steps of loading in
environmental variables in as a Raster Stack. I had originally tried
to generate candidate models using both a set of raw non-correlated environmental
variables and a set of principal components accounting for 96% of the
total variation in the environmental dataset, but the sets have to be
nested for Maxent to not output the "Warning: unused field" message