Updating variables for kuenm_mod_swd.
Reusing the same variable for batch argument led to the shell script
created to generate candidate Maxent models to be over-written. Also,
the path variable to the maxent.jar file was not interpreted correctly
as it had FALSE added to the beginning. I'm hoping that specifying the
variable for maxent.jar path at the begininning alleviates this error.