Updating kuenm-cal-eval.r.
Editing script to run kuenm with the SWD (i.e., samples with data)
format. This is necessary because samples with exact coordinates that do
not match up with the raster layers are omitted from the analysis so
extracting values for the localities I have as occurrences enables
getting around that.